+++ lib/hook_controller.dart
@@ -1,8 +1,3 @@
-/// Support for doing something awesome.
-///
-/// More dartdocs go here.
 library hook_controller;
 
 export 'src/hook_controller_base.dart';
-
-// TODO: Export any libraries intended for clients of this package.