+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 1.1.0
+
+- rename next method name to commit
+- add unregisterHook and clear methods
+- add docs
+- update description
+
 ## 1.0.0
 
 - Initial version.